+++ test/audits/reddit-analysis/opportunity-data-mapper.test.js
@@ -10,8 +10,6 @@
  * governing permissions and limitations under the License.
  */
 
-/* eslint-env mocha */
-
 import { expect } from 'chai';
 import { createOpportunityData } from '../../../src/reddit-analysis/opportunity-data-mapper.js';